testsurfer/testsurfer/requirements.txt
Django==1.11.2
django-rest==0.0.1
djangorestframework==3.6.3
gunicorn==19.7.1
pytz==2017.2
whitenoise==3.3.0
